bugfix uint8_t type on size of pointer

+++ source-sdk/utils/constants/const.hpp
@@ -2,6 +2,6 @@
 #include <cstdint>
 
 namespace utils::consts {
-	constexpr uint8_t pointerSize = sizeof(void*);
+	constexpr size_t pointerSize = sizeof(void*);
 	constexpr const char* lastLoadedModule = "serverbrowser.dll";
 }
